Add wait for transaction receipt
